WIP. Created mm2_state_machine crate.
