+++ Cargo.toml
@@ -36,6 +36,7 @@ members = [
     "mm2src/mm2_net",
     "mm2src/mm2_number",
     "mm2src/mm2_rpc",
+    "mm2src/mm2_state_machine",
     "mm2src/rpc_task",
     "mm2src/mm2_test_helpers",
     "mm2src/trezor",
