+++ mm2src/coins/Cargo.toml
@@ -69,6 +69,7 @@ mm2_metrics = { path = "../mm2_metrics" }
 mm2_net = { path = "../mm2_net" }
 mm2_number = { path = "../mm2_number"}
 mm2_rpc = { path = "../mm2_rpc" }
+mm2_state_machine = { path = "../mm2_state_machine" }
 mocktopus = "0.8.0"
 num-traits = "0.2"
 parking_lot = { version = "0.12.0", features = ["nightly"] }
