+++ mm2src/coins/tendermint/tendermint_tx_history_v2.rs
@@ -9,14 +9,14 @@ use async_trait::async_trait;
 use bitcrypto::sha256;
 use common::executor::Timer;
 use common::log;
-use common::state_machine::prelude::*;
-use common::state_machine::{StandardStateMachine, StateMachineTrait};
 use cosmrs::tendermint::abci::Code as TxCode;
 use cosmrs::tendermint::abci::Event;
 use cosmrs::tx::Fee;
 use mm2_core::mm_ctx::MmArc;
 use mm2_err_handle::prelude::MmResult;
 use mm2_number::BigDecimal;
+use mm2_state_machine::prelude::*;
+use mm2_state_machine::state_machine::StateMachineTrait;
 use primitives::hash::H256;
 use rpc::v1::types::Bytes as BytesJson;
 use std::cmp;
