+++ mm2src/mm2_net/Cargo.toml
@@ -14,8 +14,9 @@ bytes = "1.1"
 cfg-if = "1.0"
 common = { path = "../common" }
 ethkey = { git = "https://github.com/KomodoPlatform/mm2-parity-ethereum.git" }
-mm2_err_handle = { path = "../mm2_err_handle" }
 mm2_core = { path = "../mm2_core" }
+mm2_err_handle = { path = "../mm2_err_handle" }
+mm2_state_machine = { path = "../mm2_state_machine" }
 derive_more = "0.99"
 http = "0.2"
 rand = { version = "0.7", features = ["std", "small_rng", "wasm-bindgen"] }
