+++ mm2src/mm2_state_machine/src/prelude.rs
@@ -1,0 +1,3 @@
+pub use crate::state_machine::{ChangeStateExt, LastState, State, StateMachineTrait, StateResult, TransitionFrom};
+
+pub trait StandardStateMachine {}
